[FE] merge
