+++ front/src/pages/MyPage/MyPick/Content.js
@@ -1,10 +1,12 @@
 import Grid from '@mui/material/Grid';
 import styled from 'styled-components';
-import DeleteOutlinedIcon from '@mui/icons-material/DeleteOutlined';
+// import DeleteOutlinedIcon from '@mui/icons-material/DeleteOutlined';
 import Typography from '@mui/material/Typography';
-import FavoriteTwoToneIcon from '@mui/icons-material/FavoriteTwoTone';
-
+// import FavoriteTwoToneIcon from '@mui/icons-material/FavoriteTwoTone';
+import MyPickBook from './MyPickBook';
 import InfiniteScroll from 'react-infinite-scroll-component';
+// import MyPickPairing from './MyPickPairing';
+// import MyPickCollection from './MyPickCollection';
 
 const ContentContainer = styled.div`
   margin-bottom: 10rem;
@@ -35,35 +37,35 @@ const ContentContainer = styled.div`
     position: fixed;
   }
 `;
-const BookImg = styled.div`
-  .resize {
-    box-sizing: inherit;
-    width: 108px !important;
-    height: 164px !important;
-    margin-left: 10px;
-  }
-`;
+// const BookImg = styled.div`
+//   .resize {
+//     box-sizing: inherit;
+//     width: 108px !important;
+//     height: 164px !important;
+//     margin-left: 10px;
+//   }
+// `;
 
 const CommentContainer = styled.div`
   display: flex;
   flex-direction: row;
 `;
-const FlexBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-left: 20px;
-  margin-right: 10px;
-  font-size: 13px;
-  border-bottom: 1px solid #e9e9e9;
-  .comment {
-    height: 125px;
-    color: #232627;
-  }
-  .heart-star-title {
-    display: flex;
-    flex-direction: row;
-  }
-`;
+// const FlexBox = styled.div`
+//   display: flex;
+//   flex-direction: column;
+//   margin-left: 20px;
+//   margin-right: 10px;
+//   font-size: 13px;
+//   border-bottom: 1px solid #e9e9e9;
+//   .comment {
+//     height: 125px;
+//     color: #232627;
+//   }
+//   .heart-star-title {
+//     display: flex;
+//     flex-direction: row;
+//   }
+// `;
 const ButtonCSS = styled.button`
   outline: none;
   display: inline-block;
@@ -75,25 +77,32 @@ const ButtonCSS = styled.button`
   background: transparent;
 `;
 
-const Remove = styled.div`
-  color: #dee2e6;
-  font-size: 24px;
-  cursor: pointer;
-  opacity: 0;
-  &:hover {
-    color: #6741ff;
-  }
-`;
+// const Remove = styled.div`
+//   color: #dee2e6;
+//   font-size: 24px;
+//   cursor: pointer;
+//   opacity: 0;
+//   &:hover {
+//     color: #6741ff;
+//   }
+// `;
 
-const ItemContainer = styled.div`
-  &:hover {
-    ${Remove} {
-      opacity: 1;
-    }
-  }
-`;
+// const ItemContainer = styled.div`
+//   &:hover {
+//     ${Remove} {
+//       opacity: 1;
+//     }
+//   }
+// `;
 
-const Content = ({ setContent, content }) => {
+const Content = ({
+  setContent,
+  content,
+  // pairingContent,
+  // setPairingContent,
+  // collectionContent,
+  // setCollectionContent,
+}) => {
   // const [data, setData] = useState({
   //   content: content.data,
   //   hasMore: true,
@@ -101,7 +110,7 @@ const Content = ({ setContent, content }) => {
 
   // 스크롤이 바닥에 닿을때 동작하는 함수
   const fetchMoreData = () => {
-    if (content.data.length >= 100) {
+    if (content.listCount >= 100) {
       setContent({
         listCount: content.listCount,
         data: content.data,
@@ -130,12 +139,12 @@ const Content = ({ setContent, content }) => {
 
   console.log('content.data', content.data);
 
-  const onRemove = (targetId) => {
-    const newCommentList = content.data.filter(
-      (el) => el.commentId !== targetId
-    );
-    setContent({ data: newCommentList, hasMore: true });
-  };
+  // const onRemove = (targetId) => {
+  //   const newCommentList = content.data.filter(
+  //     (el) => el.commentId !== targetId
+  //   );
+  //   setContent({ data: newCommentList, hasMore: true });
+  // };
 
   const removeAll = () => {
     if (window.confirm(`모든 데이터를 정말 삭제하시겠습니까?`)) {
@@ -146,50 +155,37 @@ const Content = ({ setContent, content }) => {
   return (
     <>
       <ContentContainer>
-        <Grid container>
-          <Grid
-            item
-            xs={11}
-            sx={{
-              mt: 1,
-              mb: 1,
-              display: 'flex',
-              justifyContent: 'space-around',
-            }}
-          >
+        <Grid container style={{ backgroundColor: '#f7f4ff' }}>
+          <Grid item xs={0.5} sx={{ width: 20 }}></Grid>
+          <Grid item xs={5} sx={{ mt: 1, mb: 1 }}>
             <CommentContainer>
               <Typography
                 color="#737373"
                 sx={{
                   display: 'flex',
-                  mt: 1,
-                  mb: 1,
                 }}
                 variant="body2"
                 gutterBottom
               >
-                <input type="checkbox" name="xxx" />
-                코멘트
-                <input type="checkbox" name="xxx" />
+                <input type="checkbox" defaultChecked={true} name="xxx" />
                 페어링
-                <input type="checkbox" name="xxx" />
+                <input type="checkbox" defaultChecked={true} name="xxx" />
                 컬렉션
-                <input type="checkbox" name="xxx" />책
+                <input type="checkbox" defaultChecked={true} name="xxx" />책
               </Typography>
             </CommentContainer>
           </Grid>
 
-          <Grid item xs={1}>
+          <Grid
+            item
+            xs={6.5}
+            sx={{
+              display: 'flex',
+              flexDirection: 'row-reverse',
+            }}
+          >
             <ButtonCSS onClick={removeAll}>
-              <Typography
-                color="#737373"
-                sx={{
-                  mt: 1,
-                  mb: 1,
-                }}
-                variant="body2"
-                gutterBottom
-              >
+              <Typography color="#737373" variant="body2" gutterBottom>
                 전체 삭제
               </Typography>
             </ButtonCSS>
@@ -197,7 +193,7 @@ const Content = ({ setContent, content }) => {
         </Grid>
 
         <InfiniteScroll
-          dataLength={content.data.length}
+          dataLength={content.listCount}
           // dataLength={data.content.length}
           // next={data.content && fetchMoreData}
           next={content.data && fetchMoreData}
@@ -222,136 +218,16 @@ const Content = ({ setContent, content }) => {
           }
         >
           <div>
-            {content.data ? (
-              content.data.map((data, key) => (
-                <ItemContainer key={key}>
-                  <Grid
-                    container
-                    item
-                    xs={12}
-                    sx={{
-                      display: 'flex',
-                      flexDirection: 'row',
-                    }}
-                  >
-                    <Grid item xs={0.5} sx={{ width: 20 }}></Grid>
+            <MyPickBook content={content} setContent={setContent} />
 
-                    <Grid item xs={2}>
-                      {data && (
-                        <BookImg>
-                          <img
-                            className="resize"
-                            src={
-                              data.bookCover
-                                ? data.bookCover
-                                : '/images/cherrypick_loading.gif'
-                            }
-                            alt="book thumbnail"
-                          ></img>
-                        </BookImg>
-                      )}
-                    </Grid>
-                    <Grid item xs={9}>
-                      <FlexBox>
-                        <Typography
-                          sx={{
-                            display: 'flex',
-                            mt: 1,
-                            mb: 1,
-                            fontSize: 17,
-                            fontWeight: 400,
-                          }}
-                          variant="body2"
-                          gutterBottom
-                        >
-                          {data.title}
-                        </Typography>
-                        <Typography
-                          color="#232627"
-                          sx={{
-                            height: 125,
-                            fontWeight: 200,
-                          }}
-                          variant="body2"
-                          gutterBottom
-                        >
-                          {data.content}
-                        </Typography>
-
-                        <div className="heart-star-title">
-                          <Grid
-                            item
-                            xs={3}
-                            sx={{
-                              display: 'flex',
-
-                              alignItems: 'center',
-                            }}
-                            color="#BFBFBF"
-                          >
-                            <FavoriteTwoToneIcon
-                              sx={{ width: 19.5, height: 19.5 }}
-                              align="center"
-                              style={{ color: 'FFD8D8' }}
-                            />
-                            {data.ratingCount}
-                          </Grid>
-                          <Grid
-                            item
-                            xs={3}
-                            sx={{
-                              display: 'flex',
-
-                              alignItems: 'center',
-                            }}
-                            color="#BFBFBF"
-                          ></Grid>
-                          <Grid
-                            item
-                            xs={6}
-                            sx={{
-                              display: 'flex',
-                              flexDirection: 'row-reverse',
-                            }}
-                            align="right"
-                            color="#737373"
-                          >
-                            <div>
-                              {data.bookName ? data.bookName : null},
-                              {data.author ? data.author : null}
-                            </div>
-                          </Grid>
-                        </div>
-                      </FlexBox>
-                    </Grid>
-                    <Grid
-                      item
-                      sx={{
-                        display: 'flex',
-                        flexDirection: 'row-reverse',
-                      }}
-                    >
-                      <Remove
-                        onClick={() => {
-                          // 현재 작동 안됨 (코멘트 아이디 없음)
-                          if (
-                            window.confirm(
-                              `${data.collectionId}번째 컬렉션을 삭제하시겠습니까?`
-                            )
-                          ) {
-                            onRemove(data.collectionId);
-                          }
-                        }}
-                      >
-                        <DeleteOutlinedIcon />
-                      </Remove>
-                    </Grid>
-                  </Grid>
-                </ItemContainer>
-              ))
-            ) : (
-              <div>데이터없어용</div>
-            )}
+            {/* <MyPickPairing
+              content={pairingContent}
+              setContent={setPairingContent}
+            />
+            <MyPickCollection
+              content={collectionContent}
+              setContent={setCollectionContent}
+            /> */}
           </div>
         </InfiniteScroll>
       </ContentContainer>
