+++ front/src/pages/MyPage/MyPick/MyPick.js
@@ -1,3 +1,5 @@
+/*eslint-disable*/
+
 import Header from '../Header';
 import PageContainer from '../../../components/PageContainer';
 import Nav from '../Nav';
@@ -8,9 +10,13 @@ import styled from 'styled-components';
 import { useEffect, useState } from 'react';
 // import { asyncGetMyCommentList } from '../../../store/modules/commentSlice';
 import axios from '../../../api/axios';
-import { MY_PICK_BOOK } from '../../../api/requests';
-// import { MY_PICK_PAIRING } from '../../../api/requests';
-// import { MY_PICK_COLLECTION } from '../../../api/requests';
+import {
+  MY_PICK_BOOK,
+  MY_PICK_PAIRING,
+  MY_PICK_COLLECTION,
+  COMMENT_URL,
+} from '../../../api/requests';
+
 // 페이지네이션 처럼, 페이지네이션 요청하는 쿼리 string
 
 const Void = styled.div`
@@ -36,20 +42,46 @@ const MyPick = () => {
     data: [],
     hasMore: true,
   });
+
+  const [pairingContent, setPairingContent] = useState({
+    listCount: '',
+    data: [],
+    hasMore: true,
+  });
+
+  const [collectionContent, setCollectionContent] = useState({
+    listCount: '',
+    data: [],
+    hasMore: true,
+  });
+
   const [infiniteData, setInfiniteData] = useState({
     data: [],
     hasMore: true,
   });
 
+  /// 테스트용
+  // const fetchDataTest = async () => {
+  //   axios
+  //     .get(COMMENT_URL)
+  //     .then((response) => {
+  //       console.log('response를 알아보장', response);
+  //     })
+  //     .catch((error) => console.log('에러', error));
+  // };
+
+  // 책 북마크 데이터 가져오기
   const fetchData = async () => {
     axios
       .get(MY_PICK_BOOK)
       .then((response) => {
+        console.log('then?', response);
         setContent({
           listCount: response.data.listCount,
           data: response.data.data,
           hasMore: true,
         });
+        console.log('확인', response);
         setInfiniteData({
           content: response.data,
           hasMore: true,
@@ -58,9 +90,41 @@ const MyPick = () => {
       .catch((error) => console.log('에러', error));
   };
 
+  // 페어링 북마크 데이터 가져오기
+  const fetchPairingData = async () => {
+    axios
+      .get(MY_PICK_PAIRING)
+      .then((response) => {
+        setPairingContent({
+          listCount: response.data.listCount,
+          data: response.data.data,
+          hasMore: true,
+        });
+      })
+      .catch((error) => console.log('에러', error));
+  };
+  console.log('pairingContent 현재값', pairingContent);
+
+  // 컬렉션 북마크 데이터 가져오기
+  const fetchCollectionData = async () => {
+    axios
+      .get(MY_PICK_COLLECTION)
+      .then((response) => {
+        setCollectionContent({
+          listCount: response.data.listCount,
+          data: response.data.data,
+          hasMore: true,
+        });
+        console.log('collectionContent 현재값', collectionContent);
+      })
+      .catch((error) => console.log('에러', error));
+  };
+
   useEffect(() => {
+    // fetchDataTest();
     fetchData();
-    console.log('useEffect의 state 현재값', content);
+    fetchPairingData();
+    fetchCollectionData();
   }, []);
 
   useEffect(() => {
@@ -75,9 +139,11 @@ const MyPick = () => {
           <Nav view={view} setView={setView} content={content}></Nav>
           <Content
             content={content}
-            setInfiniteData={setInfiniteData}
-            infiniteData={infiniteData}
             setContent={setContent}
+            pairingContent={pairingContent}
+            setPairingContent={setPairingContent}
+            collectionContent={collectionContent}
+            setCollectionContent={setCollectionContent}
           ></Content>
         </Container>
       ) : (
@@ -89,7 +155,7 @@ const MyPick = () => {
               src={'/images/cherrypick_loading.gif'}
               alt="loading cherrypick"
             ></img>
-            데이터가 없어용
+            데이터가 없습니다
           </Void>
         </Container>
       )}
